move Dropdown components out of Geral subfolder
